Centralizing InMemory Employee Data for Demos
