+++ Demos/Blazorise.Demo.Material.Server/Startup.cs
@@ -1,4 +1,6 @@
 #region Using directives
+using Blazorise.Demo.Data;
+using System;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.DependencyInjection;
@@ -16,6 +18,8 @@ namespace Blazorise.Demo.Material.Server
             services
                 .AddMvc()
                 .AddNewtonsoftJson();
+            services.AddHttpClient();
+            services.AddScoped<EmployeeData>();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
