+++ Demos/Blazorise.Demo/App.razor
@@ -1,4 +1,17 @@
-﻿@*
+﻿@using System;
+@using System.Collections.Generic;
+@using System.ComponentModel.DataAnnotations;
+@using System.Linq;
+@using System.Net.Http;
+@using System.Net.Http.Json;
+@using System.Threading.Tasks;
+@using Blazorise.DataGrid;
+@using Blazorise.DataGrid.Utils;
+@using Blazorise.Demo.Models;
+@using Microsoft.AspNetCore.Components;
+@using Blazorise.Demo.Data
+
+@*
     The Router component displays whichever component has a @page
     directive matching the current URI.
 *@
@@ -64,4 +77,5 @@
         //    Color = "#ff0000",
         //},
     };
+
 }
