+++ Demos/Blazorise.Demo/Blazorise.Demo.csproj
@@ -4,6 +4,7 @@
 
   <ItemGroup>
     <PackageReference Include="Markdig" Version="0.25.0" />
+    <PackageReference Include="Microsoft.Extensions.Http" Version="5.0.0" />
   </ItemGroup>
 
   <ItemGroup>
