+++ Demos/Blazorise.Demo/Layouts/MainLayout.razor
@@ -121,7 +121,7 @@
                                         Data
                                     </BarDropdownToggle>
                                     <BarDropdownMenu>
-                                        <BarDropdownItem To="tests/datagrid/static">
+                                        <BarDropdownItem To="tests/datagrid/in-memory">
                                             In Memory
                                         </BarDropdownItem>
                                         <BarDropdownItem To="tests/datagrid/manual">
