+++ Demos/Blazorise.Demo/Models/Salary.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Blazorise.Demo.Models
+{
+    public class Salary
+    {
+        public DateTime Date { get; set; }
+        public decimal Total { get; set; }
+    }
+}
