+++ Demos/Blazorise.Demo/Pages/Tests/DataGrid/InMemoryPage.razor
@@ -1,0 +1,14 @@
+﻿@page "/tests/datagrid/in-memory"
+<Row>
+    <Column>
+        <Card Margin="Margin.Is4.OnY">
+            <CardHeader>
+                <CardTitle>Datagrid: In Memory</CardTitle>
+            </CardHeader>
+            <CardBody>
+                <DataGrid TItem="Employee">
+                </DataGrid>
+            </CardBody>
+        </Card>
+    </Column>
+</Row>
